pyproject.toml
[project]
name = "rag-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "colorama>=0.4.6",
    "faiss-cpu>=1.10.0",
    "google-cloud>=0.34.0",
    "google-cloud-vision>=3.10.1",
    "google-genai>=1.7.0",
    "google-generativeai>=0.8.4",
    "huggingface-hub>=0.29.3",
    "ipykernel>=6.29.5",
    "langchain>=0.3.20",
    "langchain-community>=0.3.19",
    "langchain-core>=0.3.44",
    "langchain-openai>=0.3.8",
    "langchain-text-splitters>=0.3.6",
    "langgraph>=0.3.9",
    "numpy>=2.2.3",
    "openai>=1.66.3",
    "pydantic>=2.10.6",
    "pypdf>=5.3.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "streamlit>=1.43.2",
    "transformers>=4.49.0",
]
